requirements.txt
# Packages maintained by thelab
versiontag==1.0.3

# Everything else
Django==1.9.6
djangorestframework==3.3.2
django-oscar==1.2.1
django-oscar-api==1.0.4
psycopg2==2.6.1
PyYAML==3.11
